mlops_project/requirements.txt
click==8.0.3
hydra-core==1.1.2
kornia==0.6.4
numpy==1.20.3
omegaconf==2.1.2
opencv_contrib_python==4.5.5.64
Pillow==9.1.0
python-dotenv==0.20.0
pytorch_lightning==1.6.1
scikit_learn==1.1.0
setuptools==58.0.4
torch==1.11.0
torchvision==0.12.0
